packages/mcp-smart-client/pyproject.toml
[project]
name = "mcp-smart-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.27.1",
    "python-dotenv>=1.2.2",
]
